general:   - fixed pointer arithmetic issues on 64-bit platforms

+++ code/audio/psp/pspchannel.cpp
@@ -18,14 +18,14 @@ PSPChannel::PSPChannel() :
     threadId(-1),
     requestRelease(false)
 {
-    unsigned int addr = (unsigned int)this;
+    uintptr_t addr = (uintptr_t)this;
 
     char str[32];
     snprintf(str, 32, "audioChannelThread%u", addr);
     this->threadId = sceKernelCreateThread(str, PSPAudioRenderer::ChannelThread, 0x12, 0x10000, 0, NULL);
     CE_ASSERT(this->threadId >= 0, "PSPChannel::PSPChannel(): sceKernelCreateThread() failed: %08x\n", this->threadId);
 
-    int error = sceKernelStartThread(this->threadId, sizeof(unsigned int), &addr);
+    int error = sceKernelStartThread(this->threadId, sizeof(uintptr_t), &addr);
     CE_ASSERT(error == 0, "PSPChannel::PSPChannel(): sceKernelStartThread() failed: %08x\n", error);
 }
